Get windows version
